lambdas/cost-tracker/requirements.txt
boto3>=1.28.0
PyYAML>=6.0
pytz>=2023.3


